trying to make it better
