+++ source/Calamari.Tests/ArgoCD/Commands/Conventions/RepositoryFactoryTests.cs
@@ -1,0 +1,10 @@
+using NUnit.Framework;
+
+namespace Calamari.Tests.ArgoCD.Commands.Conventions
+{
+    [TestFixture]
+    public class RepositoryFactoryTests
+    {
+        
+    }
+}
